Introduce the RackTestClient.
This second client allows testing ruby web apps through rack-test,
which does not require to launch a real web server. That makes the
tests faster and makes it easier to include them in a simple
`rake test` workflow.

The example is extended to demonstrate both rack-test driven and real
HTTP tests schemes.