Add support for around_each hooks.
This commit completes the hook API with an around_each method on
configuration. The method takes a block like all other hooks,
except that the block itself receives a block as last argument,
that must be called/yield for the execution to continue. This
allows executing stuff around test executions, like database
transactions or other similar resource access controls.

We do not add around_all because RSpec does not seem to support
them.

As a side effect, the various `it` blocks are removed for the
postconditions. Only one such RSpec block now exists for each
test.