Add TestCase#metadata object to track domain-specific tags.
Those tags help tracking the test context in an abstract way (such
as requester user, client application or whatever). The metadata
aims at documenting the test, and can be used in pre/post conditions
or Config#before_each to instrument the test before it is actually
ran.

A real use case shows that token-based HTTP Authentication is
easier to manage when metadata tracks the requester in an abstract
way and a before_each installs the actual token. This allows
specifying the requester manually, or through precondition
instrumentation, without having to set a concrete encrypted
Authentication HTTP header.